Started to add the Javascript for the game.
Created app.js file that holds the click handler for the game board
Included Jquery, created the function that gets called when the DOM
is loaded. Created a event handler for click events on the table cells.
Started logic for game board. Also linked the app.js file
within the index.html file.