Continued defining game logic, Added reset button. Created click handler for reset button
Created click handler for reset button that uses , defined how to win
game 8 different ways, with checkForWinner function. Created checkCell
function that uses the jquery data() method that takes the cell clicked
and gives it a data attribute of (x) if it is player 1 or (o) if it is
player 2.